requirements.txt
google-generativeai
python-dotenv
typer
rich
pytest
